-just save

+++ knapsack/brute_force.cpp
@@ -5,14 +5,14 @@
 #include "brute_force.h"
 #include <vector>
 
-int max(int a, int b) { return (a > b)? a : b; }
+int max(size_t a, size_t b) { return (a > b)? a : b; }
 // Returns the maximum value that can be put in a knapsack of capacity weight_max
-int knapsack(int weight_max, std::vector<int> &weight, std::vector<int> &value) {
+int knapsack(size_t weight_max, std::vector<size_t> &weight, std::vector<size_t> &value) {
 
-    std::vector<std::vector<int>> K (value.size() + 1, std::vector<int>(weight_max + 1));
-    for (int i = 0; i <= value.size(); ++i) {
-        for (int w = 0; w <= weight_max; w++) {
-            if (i==0 || w==0) {
+    std::vector<std::vector<size_t>> K (value.size() + 1, std::vector<size_t>(weight_max + 1));
+    for (size_t i = 0; i <= value.size(); ++i) {
+        for (size_t w = 0; w <= weight_max; w++) {
+            if (i == 0 || w == 0) {
                 K[i][w] = 0;
             } else if (weight[i - 1] <= w) {
                 K[i][w] = max(value[i - 1] + K[i - 1][w - weight[i - 1]], K[i - 1][w]);
