+++ knapsack/main.cpp
@@ -8,10 +8,10 @@
 int main () {
     std::cout<< "hi"<<std::endl;//introduce data
     //call function
-    //maybe we can create a struct to sabe all this datasets
+    //maybe we can create a struct to save the input data
     std::vector<int> value = {60, 100, 120};
     std::vector<int> weight = {10, 20, 30};
-    int  weight_max = 50;
+    size_t  weight_max = 50;
     std::cout<<"Value = " << knapsack(weight_max, weight, value) << std::endl;
 
     return 0;
